Code refactor to modularize the authoring form block
