+++ express/code/blocks/template-schema/form-data.js
@@ -1,0 +1,238 @@
+/**
+ * Form data operations - get, save, restore, and compose
+ */
+
+import { STORAGE_KEY } from './state.js';
+import { fetchPlainHtml } from './plain-html.js';
+import { showToast } from './panel.js';
+
+/**
+ * Get all form data as an object
+ */
+export function getFormData(formContainer) {
+  const data = {};
+
+  // Regular inputs
+  formContainer.querySelectorAll('.daas-input').forEach((input) => {
+    const key = input.name;
+    if (!key) return;
+
+    let value = input.value;
+
+    if (input.type === 'checkbox') {
+      value = input.checked ? 'true' : 'false';
+    } else if (input.multiple && input.tagName === 'SELECT') {
+      value = Array.from(input.selectedOptions).map((o) => o.value);
+    }
+
+    if (value) data[key] = value;
+  });
+
+  // Rich text editors
+  formContainer.querySelectorAll('.daas-rte-value').forEach((input) => {
+    if (input.name && input.value) data[input.name] = input.value;
+  });
+
+  // Multi-select values
+  formContainer.querySelectorAll('.daas-multiselect-value').forEach((input) => {
+    if (input.name && input.value) data[input.name] = input.value.split(',');
+  });
+
+  // Image dropzones
+  formContainer.querySelectorAll('.daas-dropzone[data-image-data]').forEach((dropzone) => {
+    const key = dropzone.querySelector('.daas-dropzone-input')?.name;
+    if (key) {
+      data[key] = {
+        dataUrl: dropzone.dataset.imageData,
+        fileName: dropzone.dataset.imageName,
+      };
+    }
+  });
+
+  return data;
+}
+
+/**
+ * Get saved form data from sessionStorage
+ */
+export function getSavedFormData() {
+  const storedData = JSON.parse(sessionStorage.getItem(STORAGE_KEY) || '{}');
+  return storedData.savedFormData || null;
+}
+
+/**
+ * Clear saved form data from sessionStorage
+ */
+export function clearSavedFormData() {
+  const storedData = JSON.parse(sessionStorage.getItem(STORAGE_KEY) || '{}');
+  delete storedData.savedFormData;
+  delete storedData.savedAt;
+  sessionStorage.setItem(STORAGE_KEY, JSON.stringify(storedData));
+}
+
+/**
+ * Handle save draft action - saves form data to sessionStorage
+ */
+export function handleSaveDraft(formContainer) {
+  const formData = getFormData(formContainer);
+
+  const storedData = JSON.parse(sessionStorage.getItem(STORAGE_KEY) || '{}');
+  storedData.savedFormData = formData;
+  storedData.savedAt = new Date().toISOString();
+
+  sessionStorage.setItem(STORAGE_KEY, JSON.stringify(storedData));
+  showToast('Draft saved!');
+}
+
+/**
+ * Restore form data to the form fields
+ */
+export function restoreFormData(formContainer, savedData) {
+  if (!savedData) return;
+
+  Object.entries(savedData).forEach(([key, value]) => {
+    if (typeof value === 'object' && value.dataUrl) return;
+
+    const input = formContainer.querySelector(`[name="${key}"]`);
+    if (input) {
+      if (input.type === 'checkbox') {
+        input.checked = value === 'true';
+      } else {
+        input.value = value;
+      }
+    }
+
+    // Handle Quill rich text editors
+    const rteContainer = formContainer.querySelector(`.daas-field[data-key="${key}"] .daas-rte-container`);
+    if (rteContainer) {
+      const hiddenInput = rteContainer.querySelector('.daas-rte-value');
+      if (hiddenInput) hiddenInput.value = value;
+
+      if (rteContainer.quillInstance) {
+        rteContainer.quillInstance.root.innerHTML = value;
+      } else {
+        const checkQuill = setInterval(() => {
+          if (rteContainer.quillInstance) {
+            rteContainer.quillInstance.root.innerHTML = value;
+            clearInterval(checkQuill);
+          }
+        }, 100);
+        setTimeout(() => clearInterval(checkQuill), 5000);
+      }
+    }
+
+    // Handle multi-select
+    const multiSelectValue = formContainer.querySelector(`.daas-multiselect-value[name="${key}"]`);
+    if (multiSelectValue) {
+      multiSelectValue.value = Array.isArray(value) ? value.join(',') : value;
+      const optionsPanel = multiSelectValue.closest('.daas-field').querySelector('.daas-multiselect-options');
+      if (optionsPanel) {
+        const values = Array.isArray(value) ? value : value.split(',');
+        optionsPanel.querySelectorAll('input[type="checkbox"]').forEach((cb) => {
+          cb.checked = values.includes(cb.value);
+        });
+        optionsPanel.dispatchEvent(new Event('change', { bubbles: true }));
+      }
+    }
+  });
+
+  showToast('Draft restored!');
+}
+
+/**
+ * Compose final HTML with data attributes for saving
+ */
+export async function composeFinalHtml(formData, schema) {
+  const plainHtml = await fetchPlainHtml();
+  if (!plainHtml) {
+    console.error('Could not fetch plain HTML');
+    return null;
+  }
+
+  const parser = new DOMParser();
+  const doc = parser.parseFromString(plainHtml, 'text/html');
+
+  Object.entries(formData).forEach(([key, value]) => {
+    const baseKey = key.replace(/\[\d+\]/, '[]');
+    const field = schema.fields?.find((f) => f.key === baseKey);
+
+    if (typeof value === 'object' && value.dataUrl) return;
+
+    const placeholderText = `[[${key}]]`;
+    const encodedPlaceholder = encodeURIComponent(placeholderText);
+
+    // Replace in text content
+    const walker = document.createTreeWalker(doc.body, NodeFilter.SHOW_TEXT, null, false);
+    let node;
+    while ((node = walker.nextNode())) {
+      if (node.textContent.includes(placeholderText)) {
+        const parent = node.parentElement;
+        node.textContent = node.textContent.replace(placeholderText, value || '');
+
+        if (parent && field) {
+          parent.dataset.daasKey = baseKey;
+          if (field.type) parent.dataset.daasType = field.type;
+          if (field.label) parent.dataset.daasLabel = field.label;
+          if (field.required === 'true') parent.dataset.daasRequired = 'true';
+          if (field.min) parent.dataset.daasMin = field.min;
+          if (field.max) parent.dataset.daasMax = field.max;
+        }
+      }
+    }
+
+    // Replace in href attributes
+    doc.querySelectorAll('a[href]').forEach((el) => {
+      const href = el.getAttribute('href');
+      if (href.includes(placeholderText) || href.includes(encodedPlaceholder)) {
+        const newHref = href
+          .replace(placeholderText, value || '')
+          .replace(encodedPlaceholder, value ? encodeURIComponent(value) : '');
+        el.setAttribute('href', newHref);
+      }
+    });
+
+    // Replace in alt attributes
+    doc.querySelectorAll('[alt]').forEach((el) => {
+      if (el.alt.includes(placeholderText)) {
+        el.alt = el.alt.replace(placeholderText, value || '');
+      }
+    });
+  });
+
+  return doc.body.innerHTML;
+}
+
+/**
+ * Handle create page action - composes HTML and opens in new tab
+ */
+export async function handleCreatePage(formContainer, schema) {
+  const formData = getFormData(formContainer);
+  console.log('Form data:', formData);
+
+  const finalHtml = await composeFinalHtml(formData, schema);
+  if (finalHtml) {
+    const fullHtml = `<!DOCTYPE html>
+<html>
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Preview - DaaS Generated Page</title>
+  <style>
+    body { font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif; }
+  </style>
+</head>
+<body>
+${finalHtml}
+</body>
+</html>`;
+
+    const blob = new Blob([fullHtml], { type: 'text/html' });
+    const url = URL.createObjectURL(blob);
+    window.open(url, '_blank');
+
+    showToast('Page opened in new tab!');
+  } else {
+    showToast('Failed to create page', true);
+  }
+}
+
