+++ express/code/blocks/template-schema/state.js
@@ -1,0 +1,12 @@
+/**
+ * Global state and constants for the Template Schema block
+ */
+
+export const STORAGE_KEY = 'daas-template-schema';
+
+// Global state for repeater management
+export const state = {
+  cachedPlainHtml: null,
+  repeaterCounts: {}, // { faq: 1, etc: 2 }
+};
+
